+++ restart.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-kill $(pgrep -f 'python /data/dbus-senec-enfluri/dbus-senec-enfluri.py')
+kill $(pgrep -f 'python /data/dbus-iobroker-smartmeter/dbus-iobroker-smartmeter.py')
