+++ service/run
@@ -1,3 +1,3 @@
 #!/bin/sh
 exec 2>&1
-python /data/dbus-senec-enfluri/dbus-senec-enfluri.py
+python /data/dbus-iobroker-smartmeter/dbus-iobroker-smartmeter.py
